[eas-build-job] Share device run session event schema
Move the DeviceRunSession session-events record schema, type, and NDJSON
parser out of eas-cli into @expo/eas-build-job so producers (worker,
agent-device) and consumers (eas-cli, website) share a single source of
truth for the wire format. eas-cli re-exports the type and keeps the
CLI-only display, formatting, and download logic.
