+++ packages/eas-cli/src/simulator/__tests__/events.test.ts
@@ -3,7 +3,6 @@ import {
   type DeviceRunSessionEvent,
   downloadDeviceRunSessionEventsAsync,
   formatDeviceRunSessionEvent,
-  parseDeviceRunSessionEvents,
   projectDeviceRunSessionEventsForDisplay,
 } from '../events';
 
@@ -12,39 +11,6 @@ jest.mock('../../fetch', () => ({
   ...jest.requireActual('../../fetch'),
   default: jest.fn(),
 }));
-
-describe(parseDeviceRunSessionEvents, () => {
-  it('parses, deduplicates, and sorts events from the dedicated event log', () => {
-    const first = createEvent({
-      eventId: 'first',
-      ts: '2026-07-10T12:00:00.000Z',
-    });
-    const second = createEvent({
-      eventId: 'second',
-      ts: '2026-07-10T12:00:01.000Z',
-    });
-
-    expect(
-      parseDeviceRunSessionEvents(`${second}\nplain text\n${first}\n${second}`).map(
-        ({ eventId }) => eventId
-      )
-    ).toEqual(['first', 'second']);
-  });
-
-  it('ignores malformed events', () => {
-    expect(parseDeviceRunSessionEvents(`plain text\n${JSON.stringify({ v: 1 })}`)).toEqual([]);
-  });
-
-  it('ignores events with invalid timestamps', () => {
-    expect(
-      parseDeviceRunSessionEvents(
-        `${createEvent({ eventId: 'invalid', ts: 'not-a-date' })}\n${createEvent({
-          eventId: 'valid',
-        })}`
-      ).map(({ eventId }) => eventId)
-    ).toEqual(['valid']);
-  });
-});
 
 describe(projectDeviceRunSessionEventsForDisplay, () => {
   it('collapses completed operations and carries duration onto the recorded interaction', () => {
@@ -218,10 +184,6 @@ describe(downloadDeviceRunSessionEventsAsync, () => {
     ).resolves.toEqual([]);
   });
 });
-
-function createEvent(overrides: Partial<DeviceRunSessionEvent> = {}): string {
-  return JSON.stringify(createEventValue(overrides));
-}
 
 function createEventValue(overrides: Partial<DeviceRunSessionEvent> = {}): DeviceRunSessionEvent {
   return {
